Add pagination

+++ app/src/main/java/br/egsys/pokedex/data/model/Pokemons.kt
@@ -1,7 +1,7 @@
 package br.egsys.pokedex.data.model
 
 data class Pokemons(
-    val count: Long,
+    val count: Int,
     val next: String,
     val previous: String?,
     val results: List<PokemonName>
